pre-trained random forest classifier for hourly data

hourly random forest trained to forecast hourly M4-competition data

hourly_fforms

Format

randomForest 19 entries

call

the original call to randomForest

type

classification

predicted

the predicted values of the input data based on out-of-bag sample

err.rate

vector error rates of the prediction on the input data, the i-th element being the OOB error rate for all trees up to the i-th

confusion

the confusion matrix of the prediction based on OOB data

votes

vote matrix or a matrix with one row for each input data point and one column for each class, giving the fraction or number of (OOB) votes from the random forest

oob.times

number of times cases are ‘out-of-bag’ (and thus used in computing OOB error estimate)

classes

names of all forecast-models we considered as classlabels

importance

feature importance measure

importanceSD

The “standard errors” of the permutation-based importance measure

localImportance

we didn't calculate this: hence NULL

proximity

We didn't calculate this: hence NULL

ntree

Number of trees in the forest

mtry

number of predictors sampled for spliting at each node

forest

a list that contains the entire forest

y

Responce variable

test

NULL

inbag

NULL

terms

variable attributes